Revert "Feat: loadable이용 code spliting"
This reverts commit ad6b1f60401872a3530c8ea2980bd8f5d6e988a7.
