+++ package.json
@@ -4,7 +4,6 @@
   "private": true,
   "homepage": "https://neighbors-market.netlify.app",
   "dependencies": {
-    "@loadable/component": "^5.15.2",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
