+++ src/pages/Create/components/style.jsx
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import Button from '../../../components/Button';
 
 export const HeaderTextContainer = styled.div`
   position: relative;
@@ -110,6 +111,82 @@ export const SelectedImg = styled.img`
   height: 80px;
 
   border-radius: 4px;
+`;
+export const SelectCategoryWrapper = styled.div`
+  margin-top: 15px;
+  margin-left: 16px;
+  margin-bottom: 16px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 12px 16px;
+  gap: 14px;
+  position: relative;
+
+  width: 328px;
+  height: 48px;
+
+  font-family: 'Pretendard';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 24px;
+
+  color: #212121;
+  border: 1px solid #9e9e9e;
+  border-radius: 4px;
+
+  cursor: pointer;
+  img {
+    position: absolute;
+    width: 24px;
+    top: 12px;
+    right: 16px;
+  }
+`;
+
+export const CategoryModalCover = styled.div`
+  position: absolute;
+  top: -300px;
+  left: -17px;
+  width: 360px;
+  height: ${(props) => (props.category === '식품' ? '1550px' : '1460px')};
+  background-color: ${(props) => (props.show ? 'rgba(0, 0, 0, 0.3)' : '')};
+  display: ${(props) => (props.show ? 'block' : 'none')};
+  z-index: 2;
+`;
+
+export const CategoryLabel = styled.div`
+  text-align: start;
+`;
+export const CategorySelectOptions = styled.div`
+  z-index: 2;
+
+  width: 326px;
+  position: absolute;
+  top: 48px;
+  left: 0;
+  background: #ffffff;
+  box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.3),
+    0px 4px 8px 3px rgba(0, 0, 0, 0.15);
+  border-radius: 4px;
+
+  overflow: hidden;
+  display: ${(props) => (props.show ? 'block' : 'none')};
+  padding: 8px 0;
+`;
+export const CategoryOption = styled.div`
+  padding: 14px 16px;
+  font-family: 'Pretendard';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 20px;
+
+  color: #212121;
+  &:hover {
+    background-color: #f5f5f5;
+  }
 `;
 
 export const HeadAndTextContainer = styled.div`
@@ -406,6 +483,175 @@ export const SelectPaticipantContainer = styled.div`
   .participant-control-wrapper div > span {
     min-width: 20px;
   }
+`;
+
+export const SelectDateContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-left: 16px;
+  padding-bottom: 16px;
+  font-family: 'Pretendard';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 14px;
+  line-height: 24px;
+
+  color: #212121;
+
+  position: relative;
+  margin-top: 16px;
+
+  input {
+    margin-top: 4px;
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 500;
+    width: 328px;
+    height: 48px;
+    padding: 12px 16px;
+
+    background: #f5f5f5;
+    border-radius: 4px;
+    border: 0;
+  }
+  input::placeholder {
+    color: #9e9e9e;
+  }
+`;
+
+export const MyCalendarContainer = styled.div`
+  position: relative;
+
+  z-index: 2;
+  .selectbox-wrapper {
+    text-align: start;
+    margin-left: 24px;
+    margin-bottom: 16px;
+  }
+  select {
+    border: 0;
+    height: 24px;
+
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 24px;
+
+    color: #212121;
+  }
+  .react-datepicker {
+    position: relative;
+    padding: 0 16px;
+    box-sizing: border-box;
+    border-radius: 20px 20px 0px 0px;
+    width: 330px;
+    background: #ffffff;
+    /* border: 0; */
+  }
+  .calendar_x {
+    position: absolute;
+    top: 16px;
+    left: 16px;
+    width: 24px;
+    cursor: pointer;
+  }
+  h1 {
+    text-align: center;
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 20px;
+    line-height: 24px;
+    color: #212121;
+    margin-top: 16px;
+    margin-bottom: 32px;
+  }
+  .calender-btn {
+    width: 40px;
+    height: 40px;
+    position: absolute;
+    top: 0px;
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  .calender-btn img {
+    width: 40px;
+  }
+  .button--previous {
+    left: 190px;
+  }
+  .button--next {
+    right: 20px;
+  }
+  .react-datepicker__header {
+    border: 0;
+    background-color: #ffffff;
+  }
+  .react-datepicker__month {
+    margin: 0 12px 32px 12px;
+  }
+  .react-datepicker__month-container {
+    min-height: 361px;
+  }
+
+  .react-datepicker__day-name,
+  .react-datepicker__day {
+    width: 40px;
+    height: 40px;
+
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    margin: 0;
+
+    color: #212121;
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+    border: 0;
+    outline: none;
+  }
+
+  .react-datepicker__day--disabled {
+    color: #bdbdbd;
+  }
+  .react-datepicker__day--today,
+  .react-datepicker__day--keyboard-selected {
+    background-color: #ffffff;
+    border-radius: 50%;
+  }
+
+  .react-datepicker__day--selected,
+  .react-datepicker__day--selected:hover,
+  .react-datepicker__day--selected:focus,
+  .react-datepicker__day--selected:active {
+    color: #ffffff;
+    border-radius: 50%;
+    background-color: #00d082;
+  }
+`;
+
+export const DateModalCover = styled.div`
+  position: absolute;
+  top: -1200px;
+  left: -16px;
+  width: 360px;
+  height: 1550px;
+  background-color: ${(props) => (props.show ? 'rgba(0, 0, 0, 0.3)' : '')};
+  display: ${(props) => (props.show ? 'block' : 'none')};
+  z-index: 1;
+`;
+
+export const MyCalendarButton = styled(Button)`
+  background: #00d082;
+  width: calc(330px - 32px);
+  margin-bottom: 40px;
 `;
 export const ErrorChecker = styled.img`
   position: absolute;
